% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mie-splac.r
\name{average_Mloc}
\alias{average_Mloc}
\title{average_Mloc}
\usage{
average_Mloc(wavelength, epsilon, radius, medium = 1, nmax = 10)
}
\arguments{
\item{wavelength}{real vector}

\item{epsilon}{complex vector}

\item{radius}{scalar}

\item{medium}{scalar, refractive index of surrounding medium}

\item{nmax}{truncation order}
}
\value{
data.frame with wavelength and Mloc
}
\description{
Average Mloc
}
\details{
Enhancement factor averaged over the surface
}
\author{
Baptiste Auguie
}

