\name{efficiencies}
\alias{efficiencies}
\title{efficiencies}
\usage{
  efficiencies(x, GD)
}
\arguments{
  \item{x}{real vector, size parameter}

  \item{GD}{list with Gamma, Delta, A, B}
}
\value{
  matrix of Qext, Qsca, Qabs
}
\description{
  Efficiencies
}
\details{
  Calculates the far-field efficiencies for plane-wave
  illumination
}
\author{
  Baptiste Auguie
}

