\docType{package}
\name{mie-package}
\alias{mie}
\alias{mie-package}
\title{Mie theory}
\description{
  Scattering of light by a spherical object
}
\author{
  baptiste Auguie \email{baptiste.auguie@gmail.com}
}
\references{
  Etchegoin, P. Le Ru, E., Principles of Surface-Enhanced
  Raman Spectroscopy, Elsevier, Amsterdam (2009). L.
  Novotny, E. Hecht, Principles of Nano-optics Cambridge
  University Press, 2006
}
\seealso{
  Other user: \code{\link{mie}}, \code{\link{mie_bh}}
}
\keyword{packagelibrary}

