% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation.R
\name{mie_approximation}
\alias{mie_approximation}
\title{mie_approximation}
\usage{
mie_approximation(
  radius,
  wavelength,
  epsilon,
  medium = 1.33,
  order = Inf,
  efficiency = FALSE,
  ...
)
}
\arguments{
\item{radius}{scalar}

\item{wavelength}{real vector}

\item{epsilon}{complex vector}

\item{medium}{scalar, refractive index of surrounding medium}

\item{order}{truncation order}

\item{efficiency}{logical, scale by geometrical cross-sections (unused)}

\item{...}{unused}
}
\value{
data.frame
}
\description{
Analytical approximation
}
\details{
Sphere illuminated by a plane wave
}
\seealso{
Other user: 
\code{\link{mie-package}},
\code{\link{mie_bh}()},
\code{\link{mie_ml}()},
\code{\link{mie}()}
}
\author{
Baptiste Auguie
}
\concept{user}
