% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie_layer.R
\name{mie_ml}
\alias{mie_ml}
\title{mie_ml}
\usage{
mie_ml(
  wavelength,
  epsilon,
  radii,
  n_max = 10,
  efficiency = FALSE,
  mode = c("EM", "Magnetic", "Electric"),
  order = Inf
)
}
\arguments{
\item{wavelength}{real vector}

\item{epsilon}{list of dielectric functions, from inner to outer}

\item{radii}{concentric radii of each interface, from smaller to larger}

\item{efficiency}{logical, scale by geometrical cross-sections}

\item{mode}{type of mode}

\item{order}{order of multipoles}

\item{medium}{scalar, refractive index of surrounding medium}

\item{nmax}{truncation order}
}
\value{
data.frame
}
\description{
Far-field cross-sections
}
\details{
Multilayered sphere illuminated by a plane wave
}
\examples{
library(dielectric)
library(mie)
gold <- epsAg(seq(300, 800))
a <- 30
b <- 34
c <- 35

bare <- mie(gold$wavelength, gold$epsilon, radius=a, medium=1.33, efficiency=FALSE)

leps <- list(gold$epsilon, 1.33^2, 1.5^2, 1.33^2)
# leps <- list(1.5^2, gold$epsilon,  1.33^2)
la <- list(a,b,c)
coated <- mie_ml(gold$wavelength, leps, radii=la, efficiency=FALSE)

matplot(bare$wavelength, bare[, -1], type="l", lty=1,
        xlab=expression(lambda/mu*m), ylab=expression(sigma/mu*m^2))
matlines(coated$wavelength, coated[, -1],  type="l",  lty=2)

legend("topright", c(names(bare)[-1], "coated"), col=1:3, lty=c(1,1,1,2))
}
\seealso{
Other user: 
\code{\link{mie-package}},
\code{\link{mie_approximation}()},
\code{\link{mie_bh}()},
\code{\link{mie}()}
}
\author{
Baptiste Auguie
}
\concept{user}
