\name{mie_spectrum}
\alias{mie_spectrum}
\title{mie_spectrum}
\usage{
  mie_spectrum(lambda, epsilon, radius = 0.04,
    medium = 1.333, efficiency = FALSE, order = "all")
}
\arguments{
  \item{m}{}

  \item{x}{}

  \item{u}{}
}
\value{
  data.frame
}
\description{
  mie_spectrum
}
\details{
  spectrum for a sphere
}
\author{
  Baptiste Auguie (from C. Mätzler Matlab code)
}

