% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie-splac.r
\name{psi}
\alias{psi}
\title{psi}
\usage{
psi(rho, nmax)
}
\arguments{
\item{rho}{complex vector, argument}

\item{nmax}{integer, maximum order}
}
\value{
a list with psi_n and psi'_n
}
\description{
Riccati-Bessel function psi and its derivative
}
\details{
Obtained from BesselJ, converted to spherical Bessel, and scaled
}
\author{
Baptiste Auguie
}
