% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie_layer.R
\name{susceptibilities}
\alias{susceptibilities}
\title{susceptibilities}
\usage{
susceptibilities(ls, lx, n_max)
}
\arguments{
\item{ls}{list of relative refractive index}

\item{lx}{list of size parameters}

\item{nmax}{integer, maximum order}
}
\value{
list with Gamma, Delta, A, B
}
\description{
Generalised susceptibility for multilayer Mie theory
}
\details{
Corresponds to the usual coefficients a_n, b_n, c_n, d_n
}
\author{
Baptiste Auguie
}
