\name{susceptibility}
\alias{susceptibility}
\title{susceptibility}
\usage{
  susceptibility(nmax, s, x)
}
\arguments{
  \item{nmax}{integer, maximum order}

  \item{s}{complex vector, relative refractive index}

  \item{x}{real vector, size parameter}
}
\value{
  list with Gamma, Delta, A, B
}
\description{
  Generalised susceptibility for the Mie theory
}
\details{
  Corresponds to the usual coefficients a_n, b_n, c_n, d_n
}
\author{
  Baptiste Auguie
}

