\name{xi}
\alias{xi}
\title{xi}
\usage{
  xi(rho, nmax)
}
\arguments{
  \item{rho}{complex vector, argument}

  \item{nmax}{integer, maximum order}
}
\value{
  a list with psi_n and psi'_n
}
\description{
  Riccati-Bessel function xi and its derivative
}
\details{
  Obtained from BesselH (Hankel function), converted to
  spherical Hankel, and scaled
}
\author{
  Baptiste Auguie
}

