% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie-package.r
\docType{package}
\name{mie-package}
\alias{mie-package}
\title{Mie theory}
\description{
Scattering of light by a spherical object
}
\references{
Etchegoin, P. Le Ru, E., Principles of Surface-Enhanced Raman Spectroscopy, Elsevier, Amsterdam (2009).
L. Novotny, E. Hecht, Principles of Nano-optics Cambridge University Press, 2006
}
\seealso{
Other user: 
\code{\link{mie_approximation}()},
\code{\link{mie_bh}()},
\code{\link{mie_ml}()},
\code{\link{mie}()}
}
\author{
baptiste Auguie \email{baptiste.auguie@gmail.com}
}
\concept{user}
\keyword{packagelibrary}
