% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie-splac.r
\name{mie}
\alias{mie}
\title{mie}
\usage{
mie(
  wavelength,
  epsilon,
  radius,
  medium = 1,
  n_max = ceiling(2 + max(x) + 4 * max(x)^(1/3)),
  efficiency = FALSE,
  mode = c("EM", "Magnetic", "Electric"),
  order = Inf
)
}
\arguments{
\item{wavelength}{real vector}

\item{epsilon}{complex vector}

\item{radius}{scalar}

\item{medium}{scalar, refractive index of surrounding medium}

\item{n_max}{truncation order}

\item{efficiency}{logical, scale by geometrical cross-sections}

\item{mode}{type of mode}

\item{order}{order of multipoles}
}
\value{
data.frame
}
\description{
Far-field cross-sections
}
\details{
Homogeneous sphere illuminated by a plane wave
}
\examples{
gold <- epsAu(seq(400, 800))
cross_sections <- with(gold, mie(wavelength, epsilon, radius=50, medium=1.33, efficiency=FALSE))
matplot(cross_sections$wavelength, cross_sections[, -1], type="l", lty=1,
        xlab=expression(lambda/mu*m), ylab=expression(sigma/mu*m^2))
legend("topright", names(cross_sections)[-1], col=1:3, lty=1)
}
\seealso{
Other user: 
\code{\link{mie-package}},
\code{\link{mie_approximation}()},
\code{\link{mie_bh}()},
\code{\link{mie_ml}()}
}
\author{
Baptiste Auguie
}
\concept{user}
