% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mie-splac.r
\name{susceptibility}
\alias{susceptibility}
\title{susceptibility}
\usage{
susceptibility(s, x, nmax)
}
\arguments{
\item{s}{complex vector, relative refractive index}

\item{x}{real vector, size parameter}

\item{nmax}{integer, maximum order}
}
\value{
list with Gamma, Delta, A, B
}
\description{
Generalised susceptibility for the Mie theory
}
\details{
Corresponds to the usual coefficients a_n, b_n, c_n, d_n
}
\author{
Baptiste Auguie
}

